requirements.txt
Django==6.0.2
djangorestframework==3.14.0
django-filter==24.2
drf-spectacular==0.26.1
djangorestframework-simplejwt==5.3.2
fhir.resources==7.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
